requirements.txt
requests
beautifulsoup4
python-dotenv
lxml
fake-useragent
flask
flask-cors
Flask-SQLAlchemy
APScheduler
gunicorn
